+++ .build-steps.R
@@ -20,7 +20,8 @@ remotes::install_github("r-lib/revdepcheck")
 # need to set the env variable R_BIOC_VERSION env var to 3.18.
 # per: https://github.com/r-lib/revdepcheck/issues/376
 
-Sys.setenv(R_BIOC_VERSION = 3.18)
+#Sys.setenv(R_BIOC_VERSION = 3.18)
+#revdepcheck::revdep_reset()
 revdepcheck::revdep_check(num_workers = 4)
 
 # check URLs